CPBM 2.2.0-GA

+++ citrix.cpbm.custom.model/src/main/java/com/citrix/cpbm/custom/model/TestModel.java
@@ -38,7 +38,7 @@ public class TestModel implements EntityBase {
    */
   @Id
   @GeneratedValue(strategy = GenerationType.AUTO)
-  private Long id = (Long) 0L;
+  private Long id = 0L;
 
   /**
    * creationDate.
